[modular] Separate out harness from its tests
It can create problems to distribute a tool with its tests, as the use
of the tool with certain products can cause the tests to be run in
environments incompatible with that product. For example, QEMU may
not be a supported environment for a given product, but the current GN
graph for that product will result in the scheduling of a QEMU shard that
runs the modular_test_harness_test. (Incidentally, this works today due
to the fact that infra bloats the image with more 'buildbot bundle'
packages than should be there - and it was the removal of these that
uncovered this problem.

Change-Id: I6cfb5bdbe81ef470cc81dcd1cc9e1dfde3aec793