[scenic] Multiple SessionUpdaters per FrameScheduler.
Introduces DefaultFrameScheduler::UpdateManager, a helper class that
factors out some of DefaultFrameScheduler's responsibilities, then
expands upon them.  Benefits:

1) Clearly separates update management from other scheduling tasks,
and tests it separately.  This functionality doesn't require an async
loop, making it easier to test and reason about.

2) Simplifies the SessionUpdater API; Engine now has one less method
to implement (a tricky one at that, since the invariants were subtle
and not well-documented).

3) Supports registering multiple SessionUpdaters per FrameScheduler.
This was the main motivation for this CL; factoring out the UpdateManager
makes it easier to reason about whether there is sufficient test coverage.

Testability: covered by existing and additional tests in gfx_unittests.

SCN-1394 #comment
SCN-1395 #done

Change-Id: Ib757b19e187ecf7c4b054fa9435d8914e6decdc9