[pkgfs] fix handling of partially cached packages
Amber now handles a "EEXIST" from pkgfs/install/pkg correctly, which is to
probe pkgfs/versions to see if the whole package is ready, otherwise to
continue into the needs fulfillment workflow. pkgfs is adjusted to enable
this behavior, with the bulk of effort here being adding test coverage for
this case by reintroducing a blobfs, and implementing full package install
flow.

This change re-introduces the ramdisk package, which links the ramdisk-client
library via cgo, and can host a blobfs backed by ramdisk.

The main pkgfs test is updated to use the panicerr pattern from amber, to
improve readability, and also to use the aforementioned blobfs.

Various error mappings are improved, as they were found to be missing,
causing some errors that now show up in tests to be mapped to the incorrect
value. Of most significance thinfs now passes through zx.Error.Status.

Change-Id: I1451cdf76ad44b7c9a8a581dcb3f419151fcb173