[scenic][input] Use Pixel Centers for Input
Current input events are generated from the top-left corner
of a pixel when instead we want them to be generated from
the center of a pixel. Generated rays used for hit-testing
now have a jitter of (0.5,0.5) on their origin in screen
space to account for this.

Testing: All hit-test related unit tests have been updated
to reflect this change.

SCN-1488 #done

Change-Id: I547dbe40ec0c4a85e0c53a7253d63f6cb2d366f6