[safemath][kernel] Make safemath usable from the zircon kernel
++ Enable usage of safemath from the kernel.
++ When using safemath in the kernel, restrict its visibility to
   ulib/affine for now.

Note: safemath has code paths in it which can lead to CHECK/panic
behavior.  In theory, when used "properly" this should never happen.
That said, enforcing proper usage at compile time is tricky (it's
complicated).  For now, instead of attempting to enforce/limit usage,
we simply restrict the usage of the library to just libaffine, which
is using the library in a way which is guaranteed to never panic.

ZX-4598  #comment Related change.

Change-Id: I00dcf0996eaf1c2ed40c0d98edee47950fc66a42