[identity] Add inspect to account_handler
This change adds support for basic diagnostic information using inspect:
- AccountHandler lifecycle state (uninitialized, initialized, finished)
- Account local id and active clients
- Default persona local id and active clients

Due to how tightly the different structs are tied to each other in this
crate, the testing strategy here is to check the effects on the
top-level inspector itself. Unit test on individual structs could
be added as well, but for now the coverage chosen is deliberate to
catch the spurious integration mistakes, while still avoiding
noisy duplication. For instance, the inspect `Node` being dropped
removes it from the tree, which is critical for correctness. Another
easy mistake to make is to accidentally get an off-by-one error within
the inspect hierarchy.

AUTH-210

Change-Id: I0c4d371071d9dfa005512aa1a6dd375892d1314a