[build] Fix broken Rust linking against -lzircon in Clang Zircon build
Commit 70f330b370b8914f2f8c2dbf6c261393c4f5a711 obviated this code and
it should have been removed then.  With the Clang build of Zircon, it
starts to matter to Rust links that directly link against `-lzircon`.
The excess references to the Zircon build directory preempt the
special wrapper created by //build/config/fuchsia:compiler_sysroot.
The wrapper is necessary to redirect to the right file for linking.
In the GCC build of Zircon, the wrong file happens to be acceptable
for linking too (because it's not fully stripped).  Since the Clang
build strips optimally, the wrong file can't be linked against and
Rust links against -lzircon fail.

Change-Id: I0ef2d107035a020a81aa52cf509e322de5c356e6