[scenic] Hit Testing With View Clipping
Modify the hit testing algorithm in the following ways:

1) Default node behavior is to return "true" instead of "false"
2) Only traverse down descendants if the parent node is hit.
3) Add new intersection functionality to view nodes, where the
   ray is tested agains the bounding box stored in the properties
   of the view holder.

Testing: Add several new unit tests in hittest_global_unittest.cc to
verify that a ray that does not intersect a views bounding box
does not register a hit with any geometry that exists as a child
of that view.

Also add new escher unit tests for intersections and intervals
to make sure those classes work properly.

Also updates all of the input_unittests to make sure they are
compliant with the new view-bound clipping logic.

Currently optional and toggled on with the following build flag:
scenic_enforce_view_bound_clipping=true

SCN-1471 #done

Change-Id: Ic10b0f62420405fae06871c47cd5e2a028eda145