[magma][vkext] Test that using VK_FORMAT_UNDEFINED with sysmem works
This is useful when there's another participant (e.g. a video decoder)
that actually decides on the size, and the vulkan driver just needs to
be able to deal with whatever the decoder chooses. In the future we can
change the vulkan buffer collection API to explicitly give a list of
suitable formats, but this works for now.

Change-Id: I30d9d746e56583119d40b2614b0b67949de4caac