[sdk/inspect] Deprecate garnet inspect library.
This is the first step for adding Inspect to the SDK. We must avoid
name collisions between this library and the new SDK library, so we will
be renaming the Garnet inspect implementation to "inspect_deprecated".

The original library location has been updated to simply include the
files at the new location. Once clients depend on inspect_deprecated, we
will delete garnet/public/lib/inspect.

CF-887: #comment
Change-Id: I766a7dcafa5bc9d1d1292e4529b46636c1698fa5