[bootsvc] Allow multiple calls to RootResource Get
We wish to vend the root resource to a service which can vend a
lesser-privileged VMEX resource, which we intend to expose to
trustworthy loaders and JITs.  To get there, we'd like to expose
the RootResource service to this new VmexResource service.  For that to
work, we'll need it to actually be capable of handing out duplicates of
the root resource.

Components v2 will soon give us better tools to limit access to this
service to just devcoordinator and this new "vend less privileged
resources" components, which is why we're comfortable exposing this
exceedingly-privileged service slightly more broadly.

Change-Id: Iaa630251082452cec3dbf33b3de5a88b9bc486de