[kernel] Remove memset on handle tear down
Handles require that their process_id and dispatcher be 'valid', in the
sense that they are correctly zeroed, as these members may be accessed
by dangling Handle references even whilst a Handle is free'd.

Previously the process_id was already set to 0 prior to a handle being
destroyed, although it was not asserted, but the dispatcher might be
valid all the way until the Handle destructor is invoked. As the
destructor does not zero the memory, this meant the memory needed to
be manually zeroed with a memset afterwards.

This change rearranges the Handle::TearDown code to explicitly reset
the dispatcher_ reference prior to invoking the Handle destructor, and
moves the assert on `process_id() == 0` to be prior to destructor. As
reset does zero the pointer, once the destructor has been called we now
know that all relevant fields are already zeroed and no longer need to
explicitly call memset.

Change-Id: I3c552bd7b0dec729de0c2bb107e1b995ffc5df7f