[zircon][fuzzing] Add utf_conversion fuzzer
This CL adds 'utf_conversion-fuzzer' to test the 'utf16_to_utf8' function.
The fuzzer calls ‘utf16_to_utf8’ with the input buffer, casting it to uint16_t* beforehand.
The number of code units is the size of the buffer/2, and the result is stored in a static array
Of 4MB.

Change-Id: If1b32be12847cc394cdfe36b838a3605a676aa59