[sdk] Add libtrace, libtrace-engine, libtrace-provider
libtrace-provider is exported as shared for the SDK with the name
libtrace-provider-so.so. The semi-funky name is to work around a
build issue preventing building it as libtrace-provider.so (PT-220).
trace-provider has internal C FIDL bindings that aren't (yet?)
generated at SDK client build time, thus the library can't be
included in the SDK as source.

PT-220
DX-1043 #done

Change-Id: Idfb73e5797e86002a7d9a47264257422ebe1ff4e