[ledger] inject network errors in rust cloud provider
This change adds the capacity for the rust cloud provider to inject
network errors, imitating the behavior of the C++ fake cloud provider.
To do this, we refactor the connection status: connection status is now
managed by a delegate that tells us, for each kind of request, whether
it should go through or not.

TEST=session::tests::error_injection

Change-Id: I821f2494b006a3c306a1eb61e814a5d53aa235cc