[ledger] correct the shutdown sequence in test cloud provider factory
This patch fixes the shutdown sequence in test cloud provider factory to
avoid races. In the fixed sequence we first shut down the cloud provider
process, without stopping the thread that runs services we
inject/provide into the cloud provider.

Before this change, we were running into a race where the cloud provider
was detecting that the injected service was gone, and was initiating
shutdown itself, with log entry:

```
Lost connection to the token provider, shutting down the cloud provider.
```

LE-799 #done

Test: fx shell run-test-component fuchsia-pkg://fuchsia.com/ledger_tests#meta/ledger_e2e_sync.cmx
Change-Id: I71c5a7196af6ea77e99cabf874006bbb589209b9