[botanist] Have botanist send netboot NB_BOOT repeatedly
On theory on why nothing happens after the kernel is sent is that the
target is not receiving the boot command. Send the boot command "a few"
times since there's no ACK, and it's not possible to tell if it was sent
successfully.

INTK-1200 #comment [botanist] Have botanist send netboot NB_BOOT repeatedly
DNO-647 #comment [botanist] Have botanist send netboot NB_BOOT repeatedly

Change-Id: I0727c054554ef4b60dbcccf3818bcf9f9c4ca1be