[graphics][compute]: Fix clang-format host tool path.
The definition of clang_prefix has been modified by an unrelated
CL, and this broke the build of the graphics compute tests
(which are not on the CQ yet, see [1]).

This CL fixes it and documents the acceptable values, though
a similar breakage is still possible in the future until [1]
is actually landed.

Bug: SPN-52
Test: fx build src/graphics/lib/compute:tests

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/297568

Change-Id: Ie969f2d70a1f48bd67fbbc6e874fe2a356ec3f08