[identity] Manually invoked login test
This is a manually invoked test that pulls up a token manager instance
and attempts to run login using credentials suplied by the invoker.
It consists of two pieces: a mod and a host side test runner.
The mod runs on the device under test, and is a stripped down copy
of the experimental auth_tester modified to start the authorization flow
on startup and to display a fail/success color on completion.
The host side test runner invokes the mod, drives the login and
verifies success by checking the color displayed by the mod.

The mod half will eventually be removed in favor of invoking
the login flow through some other end-to-end test that requires
Google login.

Testing: Run test with run-auth-e2e script.  Failure case verified
by running the test with the Rust impl of GoogleAuthProvider, which
has incomplete implementation.

AUTH-198 #comment

Change-Id: I221423b1cb8cda65227685263e65ddecbc004f0b