[runtests] Add test durations to summary.json
The runtests binary records duration as an integer
`duration_milliseconds` field for each test in the summary.json file.
Making assertions about the file contents in tests now requires using
regexes to match the time field, as the test durations are
nondeterministic. The alternative was to add a `clock` argument to the
`RunTest` methods so we could use dependency injection to keep test
outputs deterministic, but changing the `RunTest` method signatures
would have required changes in more places.

A lot of the diff is a result of adding a `duration_milliseconds` field
to the `Result` struct, and subsequently needing to add an argument
every time a `Result` is constructed.

Bug: IN-1266
Change-Id: I1443ac102fc61a0137c0f2598cabef2451b0d95d