[netdump] Rename `ip4` field of `iphdr` type to `ip`
The `ip4` field in the `Packet` object accepted by the filters
is of type `iphdr`, which is actually valid for both IPv4 and IPv6
packets. This change renames this field to `ip` to reduce confusion.
A constructor and reset method was also added for the `Packet` class
so that on construction, an instance is not in indeterminate state.

Bug: CONN-99
Test: Build with `--with-base //garnet/packages/tests:zircon` as a
parameter to `fx set`, then `runtests -t netdump-test` on the target.
No change to observable behavior.

Change-Id: I4b4c0e086a36d1a26c68262f7152def628c99540