[benchmarking] Remove FPS per one second window metric
As an early step towards only dividing by time attempted rendering,
remove our FPS per one second window metric.  The windowed metric makes
far less sense with the new FPS metric, and is something we'll probably
want to rethink entirely.  I don't have a replacement on hand for it,
but in practice we weren't tooling around it / monitoring it that much.

PT-172 #comment

Change-Id: Ideb488f5efafa5edd9ee785959348039eb59bb92