[protocol][clk] Add frequency scaling support for clock protocol.
This patch introduces four new methods in ZX_PROTOCOL_CLOCK and
ZX_PROTOCOL_CLOCK_IMPL. Note that ZX_PROTOCOL_CLOCK and
ZX_PROTOCOL_CLOCK_IMPL are intended to track one another 1:1 with the
only difference being that the methods in ZX_PROTOCOL_CLOCK_IMPL accept
an ID which refers to the clock being adjusted in the system and
ZX_PROTOCOL_CLOCK represents a specific clock in the system which can
be parceled out as a device for other devices to depend on.

The four new methods being introduced are as follows:

(1) IsEnabled: Determines whether or not a particular clock in the
               system is running or not.

(2) QuerySupportedRate(max_rate): Returns the highest rate supported by
                                  a clock that does not exceed
                                  `max_rate`
(3) SetRate(rate): Sets the clock rate for a particular clock to a
                   specific value.
(4) GetRate: Returns the current frequency of a particular clock.

Change-Id: Ie797fb96b306a54b369e8a11464eaadf2f63947a