[debugger] Forward error for "address of".
Using "&" in an expression to take the address of something that
resulted in an error previously ignored the error. It would instead try
to take the address of an empty value which resulted in the message
"can't take the address of a temporary" which is confusing.

This patch checks for the error and forwards it appropriately.

DX-1723

Change-Id: I7e2c61c320e112895325d41d872c49387ee37a17