[debugger] Don't have the unwinder compute names.
The AOSP unwinder will normally try to compute symbolic names of code as
it unwinds. This adds overhead and generally fails since our target
binaries are normally stripped. In our system the names are all resolved
by the client so this is completely unnecessary.

This patch turns off resolve names to increase efficiency.

TEST=None. There is no behavior change. This turns off an unused portion
of a third-party library.

Change-Id: Ia8c2c49e2e9a39cd1118d4567673626a15ed9571