[mock-function] Streamline use of matchers
This commit changes the API for custom matchers to take only the
matcher object, with the expectation that the matcher itself will hold
the expected-value state, if required.  This is easily done when using
lambdas, with the use of lambda captures.

This commit also causes the mock function to instantiate the default
argument comparison method using std::equal_to<>, only when required.

Change-Id: I4cf563e52e64358548715c8125607fb54deacb44