[netdump] Filter parser syntax logic part 1/2
This CL implements the first part of the parser syntax logic.
Specifically, `(`, `)`, `not`, `and`, `or`, and frame length filters
(unqualified `greater <len>` and `less <len>` terms) are supported.

Tests for these features required modifying the existing mock
`FilterBuilder` in `parser_test.cc` to return a `size_t` identifier
instead of a single-valued `std::monostate`. This allows identification
of mock filter construction calls by their return values, and in turn
the correct flow and ordering of filters into composition functions such
as `negation` and `conjunction` can be tested.

A follow-up CL will implement and test the rest of the parser syntax.

Bug: CONN-99
Test: Build with `--with-base //garnet/packages/tests:zircon` as a
parameter to `fx set`, then `runtests -t netdump-test` on the target.

Change-Id: If3209aa47eaff19962db9758088f792ee5f8dfbe