[wlan][iwlwifi] Remove might_sleep annotations
In Linux might_sleep is used to mark functions that can sleep to make it
easier to find cases where a sleep is used in an invalid state. For this
driver that would be sleeping while holding a spinlock. In Fuchsia we
don't use spinlocks in drivers so we don't need to worry about this
class of bugs.

Change-Id: I7d85b6ba14a0bbf2f82c37a118a4abfd59d019fc