[wlan][sme] Refactor ConnectResult enum
WLAN-1162 #comment

With this change, constructing ConnectResult::Failed variant requires
including a ConnectFailure reason (which was optional previously).

Main motivation for this change is so that we can always deduce which
step connection failure happens at based on the ConnectResult enum,
which will help with logging metric. For the same reason,
ConnectResult::BadCredentials variant is removed (though we had wanted
to remove it earlier due to privacy reason)

Test:
1. fx run-test wlan-sme-tests
2. fx run-test wlanstack-tests

Change-Id: I6823944130a922ec9520ae100b8af424cac51691