[debugger] Format function pointers in new system.
Adds support for function pointers, member function pointers, and data
member pointers to the new formatting system. This is not yet hooked up
to production code.

This new code needed to look up function names so this capability was
added to the EvalContext and the special-purpose interface in the old
formatting system was removed. This simplified various parts of the
code.

Change-Id: I4fb7b4d258728a62f5438c178f61c09220756506