[lockdep] Add Lock::AssertHeld() method.
In debug mode, this calls the underlying lock type to perform a run-time
assertion to check that the lock is held.

Importantly, this can be used to convince Clang that a lock is held
when Clang is otherwise unable to determine this statically.

For example, the code:

  object.TakeLock();  // Take the lock in a way Clang can't understand.
  object.lock.AssertHeld();
  object.guarded_value = 42;

will compile, where without the "AssertHeld()", Clang will produce
an error about writing to "guarded_value" without "mutex" being held.

The changesets:

  https://fuchsia-review.googlesource.com/c/fuchsia/+/298315
  https://fuchsia-review.googlesource.com/c/fuchsia/+/298316
  https://fuchsia-review.googlesource.com/c/fuchsia/+/298317
  https://fuchsia-review.googlesource.com/c/fuchsia/+/298318

give some examples of how this reduces how often we need to disable thread
safety analysis.

Change-Id: Iaba2c54504356529f906a8dc22ddf364002272e4