[zircon] Add Mutex::AssertHeld() method.
In debug mode, this performs a run-time check that a mutex is held,
panicking if not.

Importantly, this can be used to convince Clang that a lock is held
when Clang is otherwise unable to determine this statically.

For example, the code:

  object.TakeLock();  // Take the lock in a way Clang can't understand.
  object.mutex.AssertHeld();
  object.guarded_value = 42;

will compile, where without the "AssertHeld()", Clang will produce
an error about writing to "guarded_value" without "mutex" being held.

Change-Id: I4532fe3b7c880c2445335aa856f3af8a281fcbfa