[debugger] Swith formatting systems for "print"
The outputs of "locals" and "print" commands are changed. The old system
is still used by the stack printing code. That will be changed in a
followup.

Add some missing functionality to the new formatting system.

 - Adds a "child type" to the node so we can differentiate base classes.
 - Omits addresses for function pointers by default.
 - Don't attempt to dereference null pointers.

This regresses Rust-specific type printing somewhat. That will be
enhanced in a followup. Currently the old system uses the new system to
do this and expects things to be a certain way that makes it more
difficult to change.

Change-Id: Ia5bbee83a049c401345bb6fb3916beb24a026292