[debugger] Async acceptance of connections in DebugAgent.
This change makes accepting the connection on a background thread
instead of blocking the main one. The main driver for this change is
that the debug agent now could be interested in receiving some FIDL
messages independent of whether it's connected or not. Blocking on the
accept call would not let the message loop run, thus blocking the fidl
calls.

TEST=Currently the end-to-end connection is blocked on TC-548.
     Tested manually with `fx debug` and `fx fidlcat`.

     fx shell -- run debug_agent.cmx --port=5000 --debug-mode
     fx debug --no-agent --connect=[$(fx netaddr --fuchsia)]:5000

Change-Id: Icccb9658b95e2adcbe5c10fc6693374be678361d