[debugger] String output for new format system
Adds end-to-end function for converting a value to a formatted string.
In the first phase, the tree of FormatNodes is built up recursively.
This step is asynchronous since it may require fetching memory from the
server. In the second phase, this is converted to a string in one pass.

Converted the existing FillFormatNodeDescription to additionally fill
the node value. Currently we always want this and we can separate in the
future if needed.

Change-Id: I4971cb3227eac1dc780758960290c45aa17e3240