[fidlcat] Preferentially display base protocols.
Before this change, we would say that methods with clashing ordinals belonged to
whatever protocol was added last to the index.  With this change, we give preference to
the claim that the method belongs to a protocol that didn't have it composed in.

This has the effect that, if you call a method that is composed into some random
protocol, it won't claim that the method is part of the random protocol.  It also won't
claim that the method is part of the random protocol, if you call the method as part of
that random protocol.

DX-1731 #done

Change-Id: I2f2344ec5d617111748abe2d2cc203586c8c96ae