[debugger] Convert frame printing to new system.
Changes the frame formatting functions to use AsyncOutputBuffer. This
eliminates the need for some of the more complex async behavior.

This is the last user of the old formatting system, which is now
removed.

Add a helper function for tests to synchronously get the value of an
AsyncOutputBuffer. Update the existing tests to use this.

Change-Id: I6dbf2e6437b3fe634ea79111ee7e5421b12980e3