[intel-hda] Force tests to run on boards that should have Intel HDA.
We currently dynamically probe for the Intel HDA, and only run our test
suite if we find a suitable set of inputs / outputs. This creates the
problem that if the driver fails to come up completely, our tests won't
notice.

Add a list of boards that _should_ have the appropriate Intel HDA
hardware, and force tests to run on these boards.

Change-Id: I5cef42959bcd7c1bc8cc9d0aa936038216f08a95