[build] Separate board from bringup drivers
bringup product should be the most minimal configuration and as such
shouldn't include all drivers that are normally included by individual
board. To accomplish this, this change introduces the notion of bringup
package labels which are only used for the bringup configuration when
the board package labels will be excluded.

Bug: BLD-562
Change-Id: If40f413f8f5cedfc2def786f0c543267d21ab93a