[debugger] Continue threads individually.
Previously continuing a process or the entire system would send one
message to the backend to resume everything. This is efficient but
bypasses a lot of state on the thread. This includes clearing the stack,
sending notifications, and continuing in a specific way according to the
step mode.

This change always goes through the thread API for continuing each
thread. It is more chatty over IPC but more correct.

We keep doing the same implementation of the RemoteAPITest so ths
behavior of creating a MockRemoteAPI is now the default. Tests that did
this are now simplified.

thread_impl_test_support how has only one class to it so is now renamed.

DX-1741

Change-Id: Ibf9db4cfb81baddff7afc49af4e4343a7c7a5138