[debugger] Support tree-based formatting.
Adds the ability to format the output in an indented tree form. This is
used for the "print" and "locals" output but not for the "backtrace"
output (which needs to be more compact).

The base type eliding thresholds have been adjusted to provide much more
text in the expanded tree mode.

Change-Id: I2b4ccfcc1df6e36b1ac616975ec6c9ff06544ee4