[prebuilt] Fix internal firmware build inclusion
A bug in https://fuchsia-review.googlesource.com/c/fuchsia/+/297191
caused internal_access to always be set to false, regardless of whether
the user has internal JIRI access.  This in turn was causing even
internal builds to stop including firmware files in the build because
the firmware() GN template only includes firmware files if we have
internal JIRI access.  We should set it to true if we have access to
fuchsia_internal, as we did before the linked CL.

Tested by running zircon/scripts/download-prebuilt with both an
authorized and an unauthorized CIPD client, and observed that
zircon/prebuilt/config.gni contained the expected values for
`have_firmware`.

Bug: DNO-653 #comment Fix firmware not getting included in builds
Change-Id: I918e945ba4dd2273248a9f9cf352bbe7d9f4e8e2