[zircon][kernel][format] include local headers with "" to avoid clang-format relocating them
In some cases the headers that are tightly associated with a .cc file
are included first to ensure that the header is independently includable
elsewhere. clang-format seems to want to reorder all of the includes
except where it would cause it to mix <> and "" includes.

Switch the local includes to "" to defeat this behavior.

Fix a few minor dependency problems in a few headers.

Test: build and validate it passes CQ

Tested: build both arches and verify basic CQ functionality.
Change-Id: Iaf003548ff513b8b10000768b73dce25436fcad4