[system-update-checker] Add periodic update check
If a config file is present in the system-update-checker config-data
directory, the service will now load and use the poll_frequency_minutes
setting from it to trigger an update check at that frequency.

This change also removes the default config file included in the
amber_auto_ota config target to facilitate a later change to switch from
the amber based update checker to this one.

Bug: PKG-768 #done
Test: system-update-checker-tests
Change-Id: I72a854d1ac7082cd83b942a2491bb975e8d1db18