[magma][msd-arm-mali] Add query for total GPU time since startup
MAGMA_QUERY_TOTAL_TIME is added, which can be used to query the total
time the GPU has been active since boot. If polled this can be used to
determine recent or long-term GPU utilization.

Support for this feature is added to the Mali driver.

This query also returns the monotonic time when the utilization number was last
updated. On some GPU hardware the number may be updated only
periodically or after a round-trip to firmware, and the time difference
between two measurements may be used to calculate the GPU usage
percentage more precisely.

MA-646 #comment

Change-Id: I2fbef51d10b6af3725b79d567a8651152f07e754