[brcmfmac] Fix buffer overflow in dump functions
We weren't allocating enough space for a terminating NULL character in
Debug::PrintHexDump and Debug::PrintStringDump functions, resulting in
stack corruption.

Also, limit the number of hex values per line to 16 to make it more
readable.

The rest of the changes are refactoring to support the new test, which
simply verifies that calling the dump functions with various buffer
sizes does not crash.

TEST: included
Change-Id: Ia94b19a81b340602fcb8ffde3288284e12740917