[wlan][sme] Tweak connect cancel behavior
WLAN-1162 #comment

Suppose that a connect command arrives as a previous connect attempt is
still in progress, we observe the following behaviors in the existing
code:

1. If the previous connect attempt is still on the scan step, then we
   wait until the scan finishes (and discards the result), at which
   point we notify the user that previous connect attempt is canceled,
   and start with the new connect attempt.
2. If the previous connect attempt is already on join step or later,
   then we proceed with a join scan for the new connect attempt:
   a. If join scan succeeds and we can select a BSS, at this point we
      proceed to join step and cancel the previous connect attempt
   b. If join scan fails or we cannot select a BSS, then existing
      connection attempt continues as normal (though not sure if it
      works in practice).

With this commit, we always send out a cancel notice as soon as a new
connect command arrives. For case 1, we send out cancel notice
immediately but still wait for previous scan to finish before starting
with new join scan. This means that there may be a "down period" before
the actual connect operation starts (which is also true today). For case
2, we transition SME internal state to idle immediately unless SME is
already associated (in which case, we only transition when join scan and
BSS selection succeed).

Main motivation for this change is to simplify metric gathering. By
making each connect attempt disjoint and cancel behavior explicit,
we won't have to worry about tracking two separate connect stats at the
same time.

Test:
1. Verify manually that connect still succeeds
2. Verify manually that new connect attempt cancels an existing one
2. fx run-test wlan-sme-tests

Change-Id: I43bb68f22fff9309c130462ed8966dfb3b6a548a