[cobalt][testapp] Turn local aggregation tests back on
Changes a method of CobaltController to allow tests to
do a per-report ID check of the number of observations
generated by the EventAggregator. This makes it possible
to test assertions about the testapp project alone,
without interference from other projects which are
logging locally aggregated observations (e.g. the
cobalt_internal project).

TEST= run-test-component cobalt_testapp_no_network

Change-Id: I086b81d5786eff54d3f21104684292c43f524412