[benchmarks] Remove HandleCreate_Process test, since it doesn't work via SSH
In the SSH context, the job handle doesn't allow creating processes.
This is what we get:

$ fx push-package fuchsia_benchmarks
$ fx shell
$ /pkgfs/packages/fuchsia_benchmarks/0/test/sys/perf-test
...
[ RUN      ] HandleCreate_Process
ASSERT FAILED at (../../zircon/system/utest/perftest/handle-creation-test.cc:101): zx::process::create( *zx::job::default_job(), kName, sizeof(kName) - 1, 0, &process, &root_vmar) == ZX_OK

That fails with ZX_ERR_ACCESS_DENIED.

While we could attempt to enable this test conditionally based on
whether a suitable job handle is available, that does not seem
worthwhile.

Also fix a copy+paste error in a string.

Note that this does not affect the bots, which do not use SSH for
running this test (either for the unit-test mode or for the perf-test
mode).

The initial motivation for this change is convenience: The bots run
tests via a command that is run on boot, but there are no "fx"
commands for doing that locally, and SSH is more convenient for local
testing.  Furthermore, we expect to switch the bots to using SSH for
running perf tests at some point, so this change would become
necessary eventually.

Issue: none
Test: manual test above (otherwise, bots run this test executable in
  both unit-test mode and perf-test mode)
Change-Id: Ic5f17c4298db8ae0a0553f68559233da3018e652