[netstack3] Locally join and leave IP multicast groups
This change allows the netstack to locally keep track of what IP multicast
groups a device has joined. Note, This change does not inform other nodes
(routers) that it is interested in packets destined for a multicast group.

Tests: Unittest to make sure that we do not accept packets destined to a
multicast group until we join it, and stop accepting them after leaving.

Change-Id: I74598eb32bf31cebf90b9b04bfb5c603e6220cc0