[zircon][fit] Add static_assert for alignment requirement
Currently, the implementation of function_base always aligns
a Callable to max_align_t. This is fine for representing all primitive types,
but this assumption may be violated when dealing with 256-bit SSE values
or alignas field declarations whose alignment is >16 bytes.
this patch adds a static_assert to ensure that only types with an alignment
<= the alignment of the storage type are allowed to be stored in the function_base.

Test: Uncomment the test on line 551

Change-Id: I5cebb2b98f3bb111c38c3604273c5195b26f264d