[debugger] Show Rust struct names.
In C mode we print structs as "{ ... }", but Rust programmers expect
this to be written as "StructName{ ... }". This patch adds support for
this mode.

In minimal verbosity long template names are elided using the existing
type name eliding rules. This helps make the struct names for some
standard containers (which can stretch into hundreds of characters) much
more readable.

As with C type name eliding, the full names can be shown using "verbose
mode in the print and locals commands ("-v").

Change-Id: I0f56ed4cc78f6c78d821bfd693fbedf15b96fcec