[brcmfmac] Framework for bcdc, iovars, and sim_hw
Plumb ability to set ethernet address from driver to simulated
hardware. This introduces the following abstractions:
bcdc: parse BCDC-style messages passed from the driver proper
iovars: simulates firmware handling of variables
hardware: simulates hardware settings

Note that there's no way for us to know which features are
implemented in hardware and which are implemented in firmware, so
we will make an educated guess. We may decide at some point in
the future that the distinction is unnecessary, but for now it
makes conceptual sense to keep these distinct.

WLAN-1126

TEST: Manually smoke tested with Ubiquiti AP.
      This code is a part of the unit test infrastructure we are
      trying to bring online.
Change-Id: I4b5424142016c1b9c02c04a0fe02f6136f3316c5