[debugger] Improve Rust type printing.
Properly prints Rust enums, tuples, and tuple structs.

Converts Rust naming to use colons as in "<name>: <value>" instead of
" = " like in C. This happens for everything except toplevel variables
which keep the "=" syntax to match how local variables are set.

The recursion depth for function parameter passing in frame listing is
decreased to 3 which is a result of testing with some larger structs in
real programs which quickly become unreadable. Eliding is also more
aggressive, where it no longer prints names when values will be elided.

Set the rust flag on some symbol objects that were missed in the test
utils when it generated rust structures.

Change-Id: Ie1526ac97b7cce0be74823cb215e778ae448553f