[fidl][llcpp] Simplify Flavors #4: Move in-place flavors under InPlace.
The in-place flavor will not be as useful as before following the
introduction of the simplified flavors.
Delete the overload in SyncClient and move the static function in
Call::Foo to InPlace::Foo.

A few usages of the in-place flavor was migrated along with this CL.

TEST: No behavior change. The in-place flavor is tested by calling it
from the managed and caller-allocate flavor.

Change-Id: I91536c492f83711963fef87f51436bbbac2f06c6