[network/tests] deflake SocketTest.CloseZXSocketOnClose
This test asserts that the kernel objects (a socket and a channel) are
closed by the netstack after fuchsia.io.Node.Close is called; since
Close is called over the channel, closing the channel is asynchronous
with respect to Close returning. Thus, we need to give a reasonable
timeout to avoid flakiness in this test. Since `wait_one` returns as
soon as possible, there's no harm in making this timeout a generous 5
seconds.

Test:
  fx run-test netstack_c_api_tests -t netstack_fuchsia_c_api_test -- \
    --gtest_break_on_failure \
    --gtest_filter=SocketTest.CloseZXSocketOnClose \
    --gtest_repeat=1000000

Before this change, the test failed after ~300 iterations.

Change-Id: I4cb958325ae3e4c9721dbc62515c446b331b4be5