[netstack] avoid crash when closing sockets
This crash was introduced in an earlier refactoring, which caused
the old socket-control-plane ABI to close the socket before sending the
`Close` response on that same socket. This change undoes that, by
duplicating the socket close call; the socket-control-plane socket close
call will soon be removed as part of the removal of the socket control
plane.

NET-2364

Exempt-From-Owner-Approval: unblocking fuchsia roller

Change-Id: Id321df0257c283454a7ffd508d778824bf4965a3