[kernel] Handle::Dup should not set process_id_
Changes the handles created by Handle::Dup to not copy the process_id
of the duplicated handle, but rather start out as having an invalid
`process_id_`. Once the duplicated handle is actually added to a
`ProcessDispatcher` it will then gain its process_id.

As a duplicated handle is not actually recorded in a `ProcessDispatcher`
it is now possible to `Delete` this handle without going via
`ProcessDispatcher::RemoveHandleLocked`, where the process_id_ would be
set to ZX_KOID_INVALID. This trips an assertion in `Handle::TearDown`
that is validating the assumption that a deleted handle is not owned
by a process.

ZX-4668 #done

Change-Id: I7d5fe9c7352e48c6f34cf40ccff814f872128ae4