[zircon][sanitizer] Touch all stack pages in __sanitizer_fill_shadow test
We're seeing flake on bringup.arm64 builders. One of the working
theories is that we might stack page fault is going to increase
committed memory use eliminating the effect of __sanitizer_fill_shadow
which leads to the test failure we've observed. This change touches
all stack pages before the test to make sure this doesn't happen.

Bug: FLK-444
Change-Id: Ia791ad7e5f975be75095cda96d739699f99f7c89